requirements.txt
flask==2.0.2
gunicorn==20.1.0
python-telegram-bot==13.7
ptvsd==4.3.2 # Required for debugging.
